% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class_graph.R
\name{bGraph}
\alias{bGraph}
\title{create a branch graph}
\usage{
bGraph(n, root = "Mom", leaves = paste(l.prefix, seq(length = n), sep = ""),
  l.prefix = "D", weights = NULL, mode = c("undirected", "directed"))
}
\arguments{
\item{n}{number of leaves}

\item{root}{the root node}

\item{leaves}{the named leaves}

\item{l.prefix}{prefix to use for leaves}

\item{weights}{weights of edges to the leaves}

\item{mode}{one of "undirected" or "directed"}
}
\description{
creates a \emph{branch graph}, a simple tree with root and n branches
}
\author{
Martin Maechler
}

