% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class_graph.R
\name{class_graph}
\alias{class_graph}
\title{class graph of package}
\usage{
class_graph(package = ".")
}
\arguments{
\item{package}{the directory or package name in current environment}
}
\value{
igraph
}
\description{
class graph of package
}
\author{
Robert M Flight
}

