% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class_graph.R
\name{classTree}
\alias{classTree}
\title{generate class tree}
\usage{
classTree(Cl, where, all = FALSE)
}
\arguments{
\item{Cl}{class}

\item{where}{the package namespace to query}

\item{all}{\emph{all} of the inherited classes or just direct sub-classes to return}
}
\value{
igraph
}
\description{
Given a class name, generates the class tree for that class.
}
\details{
This function is originally written by Martin Maechler in the \code{classGraph} package.
I copied it from the \code{classGraph} package, as I did not want to depend on
\code{RGraphviz} for this package.
}
\author{
Martin Maechler
}

