% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class_graph.R
\name{get_package_env}
\alias{get_package_env}
\title{get package env}
\usage{
get_package_env(package = ".")
}
\arguments{
\item{package}{string giving the package environment to find}
}
\value{
environment
}
\description{
get package env
}
\author{
Robert M Flight
}

