% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class_graph.R
\name{multiclass_tree}
\alias{multiclass_tree}
\title{multi-class tree}
\usage{
multiclass_tree(classes, where, all = FALSE)
}
\arguments{
\item{classes}{the classes to get}

\item{where}{the package namespace of the classes}

\item{all}{get all of the sub-classes or only direct}
}
\value{
igraph
}
\description{
Given many different classes for a package, get each of their trees, and then join them
together to form a single tree
}
\author{
Robert M Flight
}

