% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class_graph.R
\name{subClasses}
\alias{subClasses}
\title{get sub-classes of a class}
\usage{
subClasses(Cl, directOnly = TRUE, complete = TRUE, ...)
}
\arguments{
\item{Cl}{class to get sub-classes of}

\item{directOnly}{get only the direct sub-classes}

\item{complete}{get complete??}
}
\description{
given a class definition, get the subclasses of that class
}
\author{
Martin Maechler
}

